+++ frontend/src/views/ImagePublish.vue
@@ -1,64 +1,19 @@
 <template>
   <div class="image-publish">
     <!-- ========== LEFT SIDEBAR ========== -->
-    <aside class="account-sidebar">
-      <div class="sidebar-header">
-        <span class="sidebar-title">账号管理</span>
-        <span class="sidebar-count">{{ totalCount }}</span>
-      </div>
-
-      <div class="group-list">
-        <div
-          v-for="group in imageAccountGroups"
-          :key="group.key"
-          :class="['group-wrap', { 'is-selected': selectedPlatform === group.key }]"
-        >
-          <!-- Group header -->
-          <div
-            class="group-header cursor-pointer"
-            @click="toggleGroup(group.key)"
-          >
-            <el-icon class="expand-icon" :style="{ color: selectedPlatform === group.key ? group.color : '' }">
-              <component :is="expandedGroups.has(group.key) ? ArrowDown : ArrowRight" />
-            </el-icon>
-            <span class="platform-badge">
-              <img v-if="group.logo" :src="group.logo" :alt="group.name" class="platform-badge-img">
-              <template v-else>{{ group.letter }}</template>
-            </span>
-            <span class="group-name">{{ group.name }}</span>
-            <span class="group-count">{{ group.accounts.filter(a => publishAccountIds.has(a.id)).length }}</span>
-          </div>
-
-          <!-- Expandable account list -->
-          <transition name="slide">
-            <div v-show="expandedGroups.has(group.key)" class="group-accounts">
-              <div
-                v-for="account in group.accounts.filter(a => publishAccountIds.has(a.id))"
-                :key="account.id"
-                :class="['account-item cursor-pointer', {
-                  active: selectedAccountId === account.id,
-                  'has-override': hasAccountOverride(account.id)
-                }]"
-                @click="selectAccount(account, group)"
-              >
-                <div class="account-avatar" :style="{ borderColor: group.color }">
-                  {{ account.name ? account.name.charAt(0) : '?' }}
-                </div>
-                <span class="account-name">{{ account.name }}</span>
-                <span :class="['dot', account.status === '正常' ? 'on' : 'off']"></span>
-                <el-icon v-if="hasAccountOverride(account.id)" class="override-icon" title="已自定义配置"><StarFilled /></el-icon>
-                <el-icon class="account-remove" @click.stop="removePublishAccount(account.id)"><Close /></el-icon>
-              </div>
-              <div v-if="group.accounts.filter(a => publishAccountIds.has(a.id)).length === 0" class="no-accounts">暂无账号</div>
-            </div>
-          </transition>
-        </div>
-      </div>
-
-      <div class="sidebar-footer">
-        <div class="add-btn cursor-pointer" @click="accountDialogVisible = true">+ 添加账号</div>
-      </div>
-    </aside>
+    <AccountSidebar
+      :account-groups="imageAccountGroups"
+      :total-count="totalCount"
+      :selected-platform="selectedPlatform"
+      :selected-account-id="selectedAccountId"
+      :expanded-groups="expandedGroups"
+      :publish-account-ids="publishAccountIds"
+      :has-account-override="hasAccountOverride"
+      @toggle-group="toggleGroup"
+      @select-account="selectAccount"
+      @remove-account="removePublishAccount"
+      @open-account-dialog="accountDialogVisible = true"
+    />
 
     <!-- ========== RIGHT MAIN AREA ========== -->
     <main class="publish-main">
@@ -280,75 +235,12 @@
     <!-- ========== DIALOGS ========== -->
 
     <!-- Account Selection Dialog -->
-    <el-dialog
+    <AccountSelectDialog
       v-model="accountDialogVisible"
-      title="选择账号"
-      width="680px"
-      :close-on-click-modal="false"
-      class="account-select-dialog"
-    >
-      <div class="account-dialog-body">
-        <div class="account-dialog-content">
-          <!-- Left: platform list (only image-capable platforms) -->
-          <div class="dialog-platform-list">
-            <div
-              :class="['dialog-platform-item', 'cursor-pointer', { active: !accountFilterPlatform }]"
-              @click="accountFilterPlatform = ''"
-            >全部平台</div>
-            <div
-              v-for="p in IMAGE_PLATFORMS"
-              :key="p.key"
-              :class="['dialog-platform-item', 'cursor-pointer', { active: accountFilterPlatform === p.name }]"
-              @click="accountFilterPlatform = p.name"
-            >
-              <span class="dialog-platform-badge">
-                <img v-if="p.logo" :src="p.logo" :alt="p.name" class="dialog-platform-badge-img">
-                <template v-else>{{ p.letter }}</template>
-              </span>
-              {{ p.name }}
-            </div>
-          </div>
-
-          <!-- Right: account checkboxes -->
-          <div class="dialog-account-list">
-            <div class="dialog-select-all">
-              <el-button size="small" @click="toggleSelectAll">
-                {{ isAllSelected ? '取消全选' : '一键全选' }}
-              </el-button>
-            </div>
-            <el-checkbox-group v-model="tempSelectedAccounts">
-              <div
-                v-for="account in filteredAccounts"
-                :key="account.id"
-                :class="['dialog-account-item', { disabled: account.status !== '正常' }]"
-              >
-                <el-checkbox :label="account.id" class="cursor-pointer">
-                  <div class="dialog-account-info">
-                    <div class="dialog-account-avatar">{{ account.name ? account.name.charAt(0) : '?' }}</div>
-                    <span class="dialog-account-name">{{ account.name }}</span>
-                    <span class="dialog-account-platform">{{ account.platform }}</span>
-                    <span :class="['dialog-account-status', account.status === '正常' ? 'ok' : 'err']">
-                      {{ account.status === '正常' ? '正常' : '已失效' }}
-                    </span>
-                  </div>
-                </el-checkbox>
-              </div>
-            </el-checkbox-group>
-            <div v-if="filteredAccounts.length === 0" class="dialog-empty">暂无可选账号</div>
-          </div>
-        </div>
-      </div>
-
-      <template #footer>
-        <div class="dialog-footer">
-          <span class="selected-count">已选择 {{ tempSelectedAccounts.length }} 个账号</span>
-          <div class="dialog-footer-btns">
-            <el-button @click="accountDialogVisible = false">取消</el-button>
-            <el-button type="primary" @click="confirmAccountSelection">确认添加</el-button>
-          </div>
-        </div>
-      </template>
-    </el-dialog>
+      :platforms="IMAGE_PLATFORMS"
+      :publish-account-ids="publishAccountIds"
+      @confirm="onAccountConfirm"
+    />
 
     <!-- Material Select Dialog -->
     <MaterialSelectDialog
@@ -365,55 +257,21 @@
     />
 
     <!-- Batch Publish Progress Dialog -->
-    <el-dialog
+    <BatchPublishDialog
       v-model="batchPublishDialogVisible"
-      title="批量发布进度"
-      width="500px"
-      :close-on-click-modal="false"
-      :close-on-press-escape="false"
-      :show-close="false"
-      class="batch-progress-dialog"
-    >
-      <div class="publish-progress">
-        <el-progress
-          :percentage="publishProgress"
-          :status="publishProgress === 100 ? 'success' : ''"
-        />
-        <div v-if="currentPublishingAccount" class="current-publishing">
-          正在发布：{{ currentPublishingAccount }}
-        </div>
-
-        <div class="publish-results" v-if="publishResults.length > 0">
-          <div
-            v-for="(result, index) in publishResults"
-            :key="index"
-            :class="['result-item', result.status]"
-          >
-            <el-icon v-if="result.status === 'success'"><Check /></el-icon>
-            <el-icon v-else-if="result.status === 'error'"><Close /></el-icon>
-            <el-icon v-else><InfoFilled /></el-icon>
-            <span class="result-label">{{ result.label }}</span>
-            <span class="result-message">{{ result.message }}</span>
-          </div>
-        </div>
-      </div>
-
-      <template #footer>
-        <div class="dialog-footer-right">
-          <el-button @click="cancelBatch" :disabled="publishProgress === 100">取消发布</el-button>
-          <el-button type="primary" @click="batchPublishDialogVisible = false" v-if="publishProgress === 100">关闭</el-button>
-        </div>
-      </template>
-    </el-dialog>
+      :progress="publishProgress"
+      :results="publishResults"
+      :current-account="currentPublishingAccount"
+      @cancel="cancelBatch"
+    />
   </div>
 </template>
 
 <script setup>
-import { ref, reactive, computed, watch, nextTick, onMounted, onBeforeUnmount } from 'vue'
+import { ref, reactive, computed, watch, nextTick, onMounted } from 'vue'
 import {
   Upload, ArrowDown, ArrowRight, Picture, PictureFilled,
-  Check, Close, InfoFilled, Promotion, StarFilled,
-  Document, WarningFilled, FullScreen
+  Promotion, Document, FullScreen
 } from '@element-plus/icons-vue'
 import { ElMessage } from 'element-plus'
 import { useAccountStore } from '@/stores/account'
@@ -422,16 +280,19 @@ import { accountApi } from '@/api/account'
 import { imagePublishApi } from '@/api/imagePublish'
 import { draftApi } from '@/api/draft'
 import { getFileUrl } from '@/utils/storage'
-import { platformList, getPlatformByKey, PLATFORMS } from '@/config/platforms'
+import { platformList, getPlatformByKey } from '@/config/platforms'
 import { useRoute } from 'vue-router'
 
+import AccountSidebar from '@/components/AccountSidebar.vue'
+import AccountSelectDialog from '@/components/AccountSelectDialog.vue'
+import BatchPublishDialog from '@/components/BatchPublishDialog.vue'
 import ImageUploader from '@/components/ImageUploader.vue'
 import ImageCarousel from '@/components/ImageCarousel.vue'
 import ImagePreviewDialog from '@/components/ImagePreviewDialog.vue'
 import MaterialSelectDialog from '@/components/MaterialSelectDialog.vue'
 import ImageCoverUpload from '@/components/ImageCoverUpload.vue'
+import { useAutoSave } from '@/composables/useAutoSave'
 
-// 渠道图文发布面板
 import DouyinImagePublishPanel from '@/components/douyin/ImagePublishPanel.vue'
 import XiaohongshuImagePublishPanel from '@/components/xiaohongshu/ImagePublishPanel.vue'
 import KuaishouImagePublishPanel from '@/components/kuaishou/ImagePublishPanel.vue'
@@ -443,7 +304,6 @@ appStore.loadAutoFillTitle()
 appStore.loadAutoSaveSettings()
 const route = useRoute()
 
-// Only image-capable platforms: 小红书(1), 抖音(3), 快手(4)
 const IMAGE_PLATFORM_KEYS = ['xiaohongshu', 'douyin', 'kuaishou']
 const IMAGE_PLATFORMS = platformList.filter(p => IMAGE_PLATFORM_KEYS.includes(p.key))
 
@@ -452,7 +312,6 @@ const expandedGroups = ref(new Set())
 const selectedPlatform = ref(null)
 const selectedAccountId = ref(null)
 
-// Account groups - only image-capable platforms
 const imageAccountGroups = computed(() => {
   return IMAGE_PLATFORMS.map(p => ({
     key: p.key,
@@ -481,13 +340,18 @@ const currentPlatformConfig = computed(() =>
   selectedPlatform.value ? getPlatformByKey(selectedPlatform.value) : null
 )
 
-// ========== Public Config (shared across all accounts) ==========
+// ========== Public Config ==========
 const commonConfig = reactive({
-  images: [],      // Array of { id, name, url, path, size, type }
-  coverImage: null, // 封面图片（公共配置）
+  images: [],
+  coverImage: null,
 })
 
 const currentPreviewIndex = ref(0)
+
+// ========== Auto-save ==========
+const currentDraftId = ref(null)
+
+const { hasChanges, startAutoSaveTimer } = useAutoSave(() => saveDraft())
 
 // ========== Channel Panel Refs & Helpers ==========
 const douyinPanelRef = ref(null)
@@ -504,12 +368,6 @@ function getAccountDisplayName(accountId) {
   return account ? account.name : '未知'
 }
 
-function getAccountName(accountId) {
-  const account = accountStore.accounts.find(a => a.id === accountId)
-  return account ? account.name : '未知'
-}
-
-// ========== Channel Panel Event Handlers ==========
 function onChannelConfigChanged() {
   hasChanges.value = true
 }
@@ -526,7 +384,7 @@ function hasAccountOverride(accountId) {
   return false
 }
 
-// ========== Batch title/description/tags sync ==========
+// ========== Batch sync ==========
 const batchTitle = ref('')
 const batchDescription = ref('')
 const batchTags = ref([])
@@ -553,7 +411,7 @@ function syncBatchToAll() {
   ElMessage.success('已同步到所有平台')
 }
 
-// ========== Init: expand first group with accounts ==========
+// ========== Init ==========
 const firstGroup = imageAccountGroups.value.find(g => g.accounts.length > 0)
 if (firstGroup) {
   expandedGroups.value.add(firstGroup.key)
@@ -563,27 +421,6 @@ if (firstGroup) {
 // ========== Dialog State ==========
 const accountDialogVisible = ref(false)
 const batchPublishDialogVisible = ref(false)
-
-// Account dialog state
-const accountFilterPlatform = ref('')
-const tempSelectedAccounts = ref([])
-
-// 账号弹窗打开时恢复状态
-watch(accountDialogVisible, async (visible) => {
-  if (visible) {
-    tempSelectedAccounts.value = [...publishAccountIds]
-    if (accountStore.accounts.length === 0) {
-      try {
-        const res = await accountApi.getAccounts()
-        if (res.code === 200 && res.data) {
-          accountStore.setAccounts(res.data)
-        }
-      } catch (e) {
-        console.error('加载账号失败:', e)
-      }
-    }
-  }
-})
 
 // Refs
 const imageUploaderRef = ref(null)
@@ -599,9 +436,6 @@ const isCancelled = ref(false)
 
 // Selected accounts for publishing
 const publishAccountIds = reactive(new Set())
-const currentDraftId = ref(null)
-const autoSaveTimer = ref(null)
-const hasChanges = ref(false)
 
 // ========== Sidebar Methods ==========
 
@@ -626,6 +460,16 @@ function selectAccount(account, group) {
   expandedGroups.value.add(group.key)
 }
 
+// ========== Account Dialog ==========
+
+function onAccountConfirm(ids) {
+  ids.forEach(id => {
+    publishAccountIds.add(id)
+  })
+  hasChanges.value = true
+  ElMessage.success(`已选择 ${ids.length} 个账号`)
+}
+
 // ========== Image Methods ==========
 
 function triggerUpload() {
@@ -643,7 +487,6 @@ function openPreviewDialog() {
 function openMaterialLibraryForImage(index) {
   materialSelectMode.value = 'image'
   materialSelectDialogRef.value?.open()
-  // Store the target index for when material is selected
   materialTargetIndex.value = index
 }
 
@@ -653,7 +496,7 @@ function openMaterialLibraryForCover() {
 }
 
 const materialTargetIndex = ref(-1)
-const materialSelectMode = ref('image') // 'image' or 'cover'
+const materialSelectMode = ref('image')
 
 function onMaterialSelected(material) {
   const imageData = {
@@ -667,7 +510,6 @@ function onMaterialSelected(material) {
     progress: 100,
   }
 
-  // 如果是选择封面
   if (materialSelectMode.value === 'cover') {
     commonConfig.coverImage = {
       id: material.id,
@@ -681,62 +523,16 @@ function onMaterialSelected(material) {
     return
   }
 
-  // 选择图片
   const targetIdx = materialTargetIndex.value
   if (targetIdx >= 0 && targetIdx < commonConfig.images.length) {
-    // Replace existing image
     commonConfig.images[targetIdx] = { ...commonConfig.images[targetIdx], ...imageData }
   } else {
-    // Add new image
     if (commonConfig.images.length < 35) {
       commonConfig.images.push(imageData)
     } else {
       ElMessage.warning('最多只能上传 35 张图片')
     }
   }
-}
-
-// ========== Account Dialog Methods ==========
-
-const filteredAccounts = computed(() => {
-  let list = accountStore.accounts
-  // Only show image-capable platform accounts
-  const imagePlatformNames = IMAGE_PLATFORMS.map(p => p.name)
-  list = list.filter(a => imagePlatformNames.includes(a.platform))
-  if (accountFilterPlatform.value) {
-    list = list.filter(a => a.platform === accountFilterPlatform.value)
-  }
-  return list
-})
-
-const validFilteredAccounts = computed(() =>
-  filteredAccounts.value.filter(a => a.status === '正常')
-)
-
-const isAllSelected = computed(() =>
-  validFilteredAccounts.value.length > 0 &&
-  validFilteredAccounts.value.every(a => tempSelectedAccounts.value.includes(a.id))
-)
-
-function toggleSelectAll() {
-  if (isAllSelected.value) {
-    const validIds = new Set(validFilteredAccounts.value.map(a => a.id))
-    tempSelectedAccounts.value = tempSelectedAccounts.value.filter(id => !validIds.has(id))
-  } else {
-    const validIds = validFilteredAccounts.value.map(a => a.id)
-    const merged = new Set([...tempSelectedAccounts.value, ...validIds])
-    tempSelectedAccounts.value = [...merged]
-  }
-}
-
-function confirmAccountSelection() {
-  tempSelectedAccounts.value.forEach(id => {
-    publishAccountIds.add(id)
-  })
-  hasChanges.value = true
-  accountDialogVisible.value = false
-  ElMessage.success(`已选择 ${tempSelectedAccounts.value.length} 个账号`)
-  tempSelectedAccounts.value = []
 }
 
 // ========== Publish Methods ==========
@@ -794,7 +590,6 @@ async function publishAll() {
     return
   }
 
-  // Delegate validation to channel components
   for (const group of imageAccountGroups.value) {
     if (group.accounts.length === 0) continue
     const panel = getPanel(group.key)
@@ -868,41 +663,6 @@ function cancelBatch() {
   ElMessage.info('正在取消发布...')
 }
 
-// ========== Auto-save ==========
-
-function startAutoSaveTimer() {
-  if (autoSaveTimer.value) {
-    clearInterval(autoSaveTimer.value)
-    autoSaveTimer.value = null
-  }
-  if (!appStore.autoSaveDraft) return
-  autoSaveTimer.value = setInterval(() => {
-    if (hasChanges.value) {
-      saveDraft()
-      hasChanges.value = false
-    }
-  }, appStore.autoSaveInterval * 1000)
-}
-
-function stopAutoSaveTimer() {
-  if (autoSaveTimer.value) {
-    clearInterval(autoSaveTimer.value)
-    autoSaveTimer.value = null
-  }
-}
-
-watch(() => appStore.autoSaveDraft, (val) => {
-  if (val) startAutoSaveTimer()
-  else stopAutoSaveTimer()
-})
-
-watch(() => appStore.autoSaveInterval, () => {
-  if (appStore.autoSaveDraft) startAutoSaveTimer()
-})
-
-// 监听内容变化
-watch(commonConfig, () => { hasChanges.value = true }, { deep: true })
-
 // ========== Old Draft Migration ==========
 function migrateOldDraftFormat(dd) {
   if (dd.commonConfig?.topics && Array.isArray(dd.commonConfig.topics)) {
@@ -970,7 +730,6 @@ async function loadDraft(draftId) {
 
     migrateOldDraftFormat(dd)
 
-    // 先恢复 UI 状态和发布账号
     if (dd.selectedPlatform) selectedPlatform.value = dd.selectedPlatform
     if (dd.selectedAccountId) {
       selectedAccountId.value = dd.selectedAccountId
@@ -983,14 +742,12 @@ async function loadDraft(draftId) {
       dd.publishAccountIds.forEach(id => publishAccountIds.add(id))
     }
 
-    // 等 Vue 将 selectedAccountId prop 传递到子组件后再还原配置
     await nextTick()
 
     if (dd.platformConfigs) {
       for (const [key, val] of Object.entries(dd.platformConfigs)) {
         const panel = getPanel(key)
         if (panel && val) {
-          // 只传属于本平台的账号覆盖，不同渠道互不干扰
           const ownOverrides = {}
           if (dd.accountOverrides) {
             const ownAccountIds = new Set(
@@ -1021,10 +778,12 @@ function getPlatformKeyByName(platformName) {
   return platform?.key || ''
 }
 
+// Watch content changes
+watch(commonConfig, () => { hasChanges.value = true }, { deep: true })
+
 onMounted(async () => {
   startAutoSaveTimer()
 
-  // 加载账号列表，确保侧栏和草稿还原能正常工作
   if (accountStore.accounts.length === 0) {
     try {
       const res = await accountApi.getAccounts()
@@ -1036,15 +795,10 @@ onMounted(async () => {
     }
   }
 
-  // 检查是否有 draft 参数
   const draftId = route.query.draft
   if (draftId) {
     await loadDraft(Number(draftId))
   }
-})
-
-onBeforeUnmount(() => {
-  stopAutoSaveTimer()
 })
 </script>
 
@@ -1060,241 +814,6 @@ onBeforeUnmount(() => {
   height: 100%;
   gap: 0;
   overflow: hidden;
-}
-
-// ========== LEFT SIDEBAR ==========
-.account-sidebar {
-  width: 232px;
-  flex-shrink: 0;
-  background: linear-gradient(180deg, #0d0d22 0%, #0a0a1a 100%);
-  border-right: 1px solid rgba(255, 255, 255, 0.06);
-  display: flex;
-  flex-direction: column;
-  overflow: hidden;
-
-  .sidebar-header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 20px 16px 16px;
-
-    .sidebar-title {
-      font-size: 16px;
-      font-weight: 700;
-      color: #e2e8f0;
-    }
-
-    .sidebar-count {
-      font-size: 11px;
-      color: #a78bfa;
-      background: rgba(139, 92, 246, 0.12);
-      padding: 3px 10px;
-      border-radius: 12px;
-      font-weight: 700;
-    }
-  }
-
-  .group-list {
-    flex: 1;
-    overflow-y: auto;
-    padding: 4px 0;
-
-    &::-webkit-scrollbar { width: 3px; }
-    &::-webkit-scrollbar-thumb { background: rgba(139, 92, 246, 0.15); border-radius: 2px; }
-  }
-
-  .group-wrap {
-    margin: 2px 10px;
-    border-radius: 10px;
-    transition: all 0.2s ease;
-    border: 1px solid transparent;
-
-    &.is-selected {
-      background: rgba(139, 92, 246, 0.1);
-      border-color: rgba(139, 92, 246, 0.2);
-      margin: 2px 9px;
-    }
-  }
-
-  .group-header {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-    padding: 11px 12px;
-    border-radius: 10px;
-    transition: all 0.2s ease;
-    user-select: none;
-
-    &:hover { background: rgba(255, 255, 255, 0.03); }
-
-    .expand-icon {
-      font-size: 12px;
-      color: $text-muted;
-      transition: all 0.2s ease;
-    }
-
-    .platform-badge {
-      width: 34px;
-      height: 34px;
-      border-radius: 10px;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      color: #fff;
-      font-size: 14px;
-      font-weight: 700;
-      flex-shrink: 0;
-      overflow: hidden;
-      box-shadow: 0 2px 8px rgba(0, 0, 0, 0.4);
-
-      .platform-badge-img {
-        width: 24px;
-        height: 24px;
-        object-fit: contain;
-      }
-    }
-
-    .group-name {
-      flex: 1;
-      font-size: 15px;
-      color: $text-secondary;
-      font-weight: 600;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-    }
-
-    .group-count {
-      font-size: 11px;
-      color: #a78bfa;
-      background: rgba(139, 92, 246, 0.1);
-      padding: 2px 8px;
-      border-radius: 8px;
-      font-weight: 600;
-    }
-  }
-
-  .group-accounts {
-    padding: 0 12px 8px 48px;
-
-    .no-accounts {
-      font-size: 12px;
-      color: $text-muted;
-      padding: 6px 0;
-    }
-  }
-
-  // Slide transition
-  .slide-enter-active, .slide-leave-active {
-    transition: all 200ms ease;
-    overflow: hidden;
-  }
-  .slide-enter-from, .slide-leave-to { opacity: 0; max-height: 0; }
-  .slide-enter-to, .slide-leave-from { opacity: 1; max-height: 500px; }
-
-  .account-item {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    padding: 7px 8px;
-    border-radius: 8px;
-    transition: all 0.2s ease;
-    border: 1px solid transparent;
-
-    &:hover {
-      background: rgba(255, 255, 255, 0.04);
-      border-color: rgba(255, 255, 255, 0.04);
-    }
-
-    &.active {
-      background: rgba(139, 92, 246, 0.1);
-      border-color: rgba(139, 92, 246, 0.15);
-    }
-
-    .account-avatar {
-      width: 24px;
-      height: 24px;
-      border-radius: 50%;
-      background: rgba(139, 92, 246, 0.15);
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      font-size: 10px;
-      color: #c4b5fd;
-      font-weight: 700;
-      flex-shrink: 0;
-      border: 2px solid transparent;
-      transition: all 0.2s ease;
-    }
-
-    .account-name {
-      flex: 1;
-      font-size: 12px;
-      color: $text-secondary;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-      font-weight: 500;
-    }
-
-    .dot {
-      width: 7px;
-      height: 7px;
-      border-radius: 50%;
-      flex-shrink: 0;
-
-      &.on { background: $success-color; box-shadow: 0 0 6px rgba(34, 197, 94, 0.5); }
-      &.off { background: $danger-color; box-shadow: 0 0 6px rgba(239, 68, 68, 0.5); }
-    }
-
-    .account-remove {
-      font-size: 16px;
-      color: $text-muted;
-      opacity: 0;
-      transition: all 0.15s ease;
-      flex-shrink: 0;
-      margin-left: 4px;
-      cursor: pointer;
-
-      &:hover { color: $danger-color; opacity: 1 !important; }
-    }
-
-    &:hover .account-remove { opacity: 0.5; }
-
-    &.has-override {
-      background: rgba(245, 158, 11, 0.06);
-      border-color: rgba(245, 158, 11, 0.1);
-      .account-name { font-weight: 600; }
-    }
-
-    .override-icon {
-      font-size: 12px;
-      color: #f59e0b;
-      flex-shrink: 0;
-    }
-  }
-
-  .sidebar-footer {
-    padding: 12px 10px;
-    border-top: 1px solid rgba(255, 255, 255, 0.04);
-
-    .add-btn {
-      border: 1.5px dashed rgba(139, 92, 246, 0.25);
-      border-radius: 10px;
-      padding: 10px;
-      text-align: center;
-      font-size: 13px;
-      font-weight: 600;
-      color: #a78bfa;
-      transition: all 0.2s ease;
-
-      &:hover {
-        border-color: rgba(139, 92, 246, 0.5);
-        color: #c4b5fd;
-        background: rgba(139, 92, 246, 0.08);
-      }
-    }
-  }
 }
 
 // ========== RIGHT MAIN ==========
@@ -1442,7 +961,6 @@ onBeforeUnmount(() => {
   50% { border-color: rgba(255, 120, 117, 0.5); box-shadow: 0 0 20px rgba(255, 77, 79, 0.12); }
 }
 
-// ========== Section Bar — big gradient accent ==========
 .section-bar {
   display: flex;
   align-items: center;
@@ -1476,12 +994,10 @@ onBeforeUnmount(() => {
   }
 }
 
-// ========== Cover Section ==========
 .cover-section {
   margin-bottom: 16px;
 }
 
-// ========== Media Section ==========
 .media-section {
   margin-bottom: 20px;
   border: 1px solid rgba(139, 92, 246, 0.12);
@@ -1495,7 +1011,6 @@ onBeforeUnmount(() => {
   }
 }
 
-// ========== Form Fields ==========
 .form-field {
   margin-bottom: 20px;
 
@@ -1531,14 +1046,12 @@ onBeforeUnmount(() => {
   }
 }
 
-// ========== Divider — fade ==========
 .divider {
   height: 1px;
   background: linear-gradient(90deg, transparent, rgba(139, 92, 246, 0.15) 30%, rgba(139, 92, 246, 0.15) 70%, transparent);
   margin: 8px 0 28px;
 }
 
-// ========== Batch Sync ==========
 .batch-sync-section {
   border: 1px solid rgba(255, 255, 255, 0.06);
   border-radius: 14px;
@@ -1572,7 +1085,6 @@ onBeforeUnmount(() => {
   }
 }
 
-// ========== Platform Title & Description ==========
 .platform-title-desc {
   display: flex;
   flex-direction: column;
@@ -1580,15 +1092,12 @@ onBeforeUnmount(() => {
   margin-bottom: 12px;
 }
 
-// ========== Settings Grid ==========
 .settings-grid {
   display: grid;
   grid-template-columns: repeat(auto-fill, minmax(280px, 1fr));
   gap: 12px;
 }
 
-
-// ========== Setting Card ==========
 .setting-card {
   padding: 16px 18px;
   border: 1px solid;
@@ -1675,7 +1184,6 @@ onBeforeUnmount(() => {
   }
 }
 
-// ========== Hotspot Tags ==========
 .hotspot-tags-list {
   display: flex;
   flex-wrap: wrap;
@@ -1698,7 +1206,6 @@ onBeforeUnmount(() => {
   margin-bottom: 8px;
 }
 
-// ========== No Platform Hint ==========
 .no-platform-hint {
   display: flex;
   flex-direction: column;
@@ -1865,7 +1372,6 @@ onBeforeUnmount(() => {
   margin-left: 8px;
 }
 
-// ========== Cover Action Buttons ==========
 .cover-action-btn {
   display: flex;
   align-items: center;
@@ -1908,175 +1414,6 @@ onBeforeUnmount(() => {
   }
 }
 
-// ========== Account Dialog ==========
-.account-select-dialog {
-  .account-dialog-body {
-    .account-dialog-content {
-      display: flex;
-      gap: 0;
-      border: 1px solid rgba(255, 255, 255, 0.06);
-      border-radius: 12px;
-      overflow: hidden;
-      height: 420px;
-    }
-
-    .dialog-platform-list {
-      width: 160px;
-      flex-shrink: 0;
-      border-right: 1px solid rgba(255, 255, 255, 0.06);
-      background: rgba(0, 0, 0, 0.25);
-      overflow-y: auto;
-
-      .dialog-platform-item {
-        padding: 14px 16px;
-        font-size: 15px;
-        color: $text-secondary;
-        display: flex;
-        align-items: center;
-        gap: 12px;
-        transition: all 0.2s ease;
-        border-left: 3px solid transparent;
-
-        &:hover { background: rgba(255, 255, 255, 0.03); }
-
-        &.active {
-          background: rgba(139, 92, 246, 0.08);
-          color: #f8fafc;
-          font-weight: 600;
-          border-left-color: $brand-start;
-        }
-
-        .dialog-platform-badge {
-          width: 28px;
-          height: 28px;
-          border-radius: 6px;
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          color: #fff;
-          font-size: 11px;
-          font-weight: 700;
-          flex-shrink: 0;
-          overflow: hidden;
-
-          .dialog-platform-badge-img { width: 22px; height: 22px; object-fit: contain; }
-        }
-      }
-    }
-
-    .dialog-account-list {
-      flex: 1;
-      padding: 12px;
-      overflow-y: auto;
-
-      .dialog-select-all {
-        display: flex;
-        justify-content: flex-end;
-        margin-bottom: 8px;
-        padding-bottom: 8px;
-        border-bottom: 1px solid rgba(255, 255, 255, 0.04);
-      }
-
-      .dialog-account-item {
-        padding: 8px 10px;
-        border-radius: 8px;
-        transition: all 0.2s ease;
-        margin-bottom: 2px;
-
-        &:hover { background: rgba(255, 255, 255, 0.03); }
-        &.disabled { opacity: 0.5; }
-      }
-
-      .dialog-account-info {
-        display: flex;
-        align-items: center;
-        gap: 8px;
-
-        .dialog-account-avatar {
-          width: 26px;
-          height: 26px;
-          border-radius: 50%;
-          background: rgba(139, 92, 246, 0.12);
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          font-size: 11px;
-          color: #c4b5fd;
-          font-weight: 700;
-          flex-shrink: 0;
-        }
-
-        .dialog-account-name { font-size: 13px; color: #f8fafc; font-weight: 500; }
-        .dialog-account-platform { font-size: 11px; color: $text-muted; }
-
-        .dialog-account-status {
-          font-size: 11px;
-          margin-left: auto;
-          &.ok { color: $success-color; }
-          &.err { color: $danger-color; }
-        }
-      }
-    }
-  }
-
-  .dialog-footer {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-
-    .selected-count { font-size: 13px; color: $text-muted; }
-    .dialog-footer-btns { display: flex; gap: 8px; }
-  }
-}
-
-// ========== Batch Progress Dialog ==========
-.batch-progress-dialog {
-  .publish-progress {
-    padding: 12px 0;
-
-    .current-publishing {
-      margin: 16px 0;
-      text-align: center;
-      color: $text-secondary;
-      font-size: 14px;
-    }
-
-    .publish-results {
-      margin-top: 20px;
-      border-top: 1px solid rgba(255, 255, 255, 0.06);
-      padding-top: 16px;
-      max-height: 300px;
-      overflow-y: auto;
-
-      .result-item {
-        display: flex;
-        align-items: center;
-        padding: 8px 0;
-        color: $text-secondary;
-
-        .el-icon { margin-right: 8px; }
-        .result-label { margin-right: 10px; font-weight: 600; color: #f8fafc; }
-        .result-message { color: $text-muted; font-size: 13px; }
-
-        &.success { .el-icon, .result-label { color: $success-color; } }
-        &.error { .el-icon, .result-label { color: $danger-color; } }
-        &.cancelled { color: $text-muted; .result-label { color: $text-muted; } }
-      }
-    }
-  }
-}
-
-// ========== Shared ==========
-.dialog-footer-right { display: flex; justify-content: flex-end; gap: 8px; }
-
-.dialog-empty {
-  text-align: center;
-  padding: 40px 0;
-  color: $text-muted;
-  font-size: 14px;
-}
-
-// ========== Douyin Styles ==========
 .setting-hint {
   font-size: 12px;
   color: $text-muted;
